+++ packages/cozy-client/src/helpers.js
@@ -1,13 +1,16 @@
 import { Association } from './associations'
 
 export const dehydrate = document => {
-  const dehydrated = {}
-  Object.entries(document).forEach(([key, value]) => {
-    if (!(value instanceof Association)) {
-      dehydrated[key] = value
-    } else if (value.raw) {
-      dehydrated[key] = value.raw
-    }
-  })
+  const dehydrated = Object.entries(document).reduce(
+    (document, [key, value]) => {
+      if (!(value instanceof Association)) {
+        document[key] = value
+      } else if (value.dehydrate) {
+        document = value.dehydrate(document)
+      }
+      return document
+    },
+    {}
+  )
   return dehydrated
 }