+++ CHANGELOG.md
@@ -3,7 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-      <a name="5.5.0"></a>
+       <a name="5.6.0"></a>
+# [5.6.0](https://github.com/cozy/cozy-client/compare/v5.5.0...v5.6.0) (2019-01-17)
+
+
+### Features
+
+* Dehydrate more relationships ([f20b7f9](https://github.com/cozy/cozy-client/commit/f20b7f9))
+* New HasMany relations API ([7b11927](https://github.com/cozy/cozy-client/commit/7b11927))
+
+
+
+
+       <a name="5.5.0"></a>
 # [5.5.0](https://github.com/cozy/cozy-client/compare/v5.4.6...v5.5.0) (2019-01-15)
 
 
@@ -14,7 +26,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-      <a name="5.4.6"></a>
+     <a name="5.4.6"></a>
 ## [5.4.6](https://github.com/cozy/cozy-client/compare/v5.4.5...v5.4.6) (2019-01-14)
 
 
@@ -25,7 +37,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-   <a name="5.4.5"></a>
+  <a name="5.4.5"></a>
 ## [5.4.5](https://github.com/cozy/cozy-client/compare/v5.4.3...v5.4.5) (2019-01-11)
 
 
@@ -33,7 +45,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package undefined
 
- <a name="5.4.4"></a>
+<a name="5.4.4"></a>
 ## [5.4.4](https://github.com/cozy/cozy-client/compare/v5.4.3...v5.4.4) (2019-01-11)